04.Python and PostgreSQL/requirements.txt
astroid==2.0.4
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.3
djangorestframework==3.9.0
entrypoints==0.2.3
factory-boy==2.11.1
Faker==0.9.2
httpie==1.0.0
idna==2.7
ipdb==0.11
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
pandocfilters==1.4.2
parso==0.3.1
pep8==1.7.1
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==2.0.7
psycopg2==2.8.2
Pygments==2.2.0
pylint==2.1.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-dateutil==2.7.5
python-dotenv==0.10.2
pytz==2018.7
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.1
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.10.11
